Temporarily suppress mobile app error (CANVAS-261)
Version 3.14.2 of the iOS Canvas app appears to create a TYPE_PUSH
communication channel if the user has push notifications enabled. This
feature requires AWS SNS to be configured, so the user gets an error
upon app launch/login.

This fix suppresses the error by returning an expected response status.